Wizard die Klassennamen wurde geändert!
